requirements.txt
aiogram
aiohttp
aiohttp_cors
firebase-admin
python-dotenv